feat: Add loading indicators to forms
Implemented loading indicators for the callback form, contact form, and newsletter subscription form. This improves user experience by providing visual feedback during form submission.

The loading indicators are displayed using a custom component (`LoadingButtonSpinner`) that utilizes the `eos-icons:loading` icon. This component is displayed while the form submission is in progress, and disappears once the submission is complete.

This commit also includes updates to the toast messages to provide more informative feedback to the user.